refactor: 메서드명 변경 및 LikeInfo dto 패키지로 이동
