+++ backend/src/main/java/hanglog/like/dto/LikeElement.java
@@ -1,6 +1,5 @@
 package hanglog.like.dto;
 
-import hanglog.like.domain.LikeInfo;
 import java.util.Collections;
 import java.util.Set;
 import lombok.AllArgsConstructor;
@@ -16,10 +15,6 @@ public class LikeElement {
 
     public boolean isLike(final Long memberId) {
         return memberIds.contains(memberId);
-    }
-
-    public LikeInfo toLikeInfo(final Long memberId) {
-        return new LikeInfo(this.getLikeCount(), this.isLike(memberId));
     }
 
     public static LikeElement empty(final Long tripId) {
