+++ backend/src/main/java/hanglog/like/dto/LikeInfo.java
@@ -1,4 +1,4 @@
-package hanglog.like.domain;
+package hanglog.like.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
